# driverlocationAPI/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The API provides finding the nearest driver location with a given GeoJSON point and radius.",
        "title": "Driver Location API",
        "contact": {
            "name": "Mert Özler",
            "url": "http://github.com/mertozler",
            "email": "meozler@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7000",
    "basePath": "/",
    "paths": {
        "/createDriver": {
            "post": {
                "description": "An endpoint for creating a driver location. It would support batch operations to handle the bulk update.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver-Create"
                ],
                "summary": "Create a Driver Location",
                "parameters": [
                    {
                        "description": "Driver data",
                        "name": "driver",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RiderCreateAndUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.DriverResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.DriverResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.DriverResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.DriverResponse"
                        }
                    }
                }
            }
        },
        "/findDriver": {
            "post": {
                "description": "The endpoint that allows searching with a GeoJSON point to find a driver if it matches the given criteria. Otherwise, the service should respond with a 404 - Not Found",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver-Find"
                ],
                "summary": "Find The Driver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Apikey header: (You need to add Bearer to the beginning of the jwt key. For example: Bearer apiKey)",
                        "name": "Apikey",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Rider data",
                        "name": "rider",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RiderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.DriverResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.DriverResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.DriverResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.DriverResponse"
                        }
                    }
                }
            }
        },
        "/getAllDriversLocation": {
            "get": {
                "description": "Returns the current location of all drives.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drivers-Location-Get-All"
                ],
                "summary": "Get All Driver Location",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.DriverResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.DriverResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.DriverResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.DriverResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "fiber.Map": {
            "type": "object",
            "additionalProperties": true
        },
        "models.Location": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "coordinates": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.RiderCreateAndUpdateRequest": {
            "type": "object",
            "required": [
                "location"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/models.Location"
                }
            }
        },
        "models.RiderRequest": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "coordinates": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "id": {
                    "type": "string"
                },
                "radius": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "responses.DriverResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/fiber.Map"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}